requirements.txt
# GUI Framework
PySide6>=6.6.0

# File Processing
python-docx>=1.1.0
pypdf>=3.17.0
pdfplumber>=0.10.3
Pillow>=10.2.0
pytesseract>=0.3.10

# API and HTTP
requests>=2.31.0
httpx>=0.26.0
pydantic>=2.5.0

# NLP and Text Processing
scikit-learn>=1.4.0
nltk>=3.8.1

# Data Handling
python-dateutil>=2.8.2

# Development and Testing
pytest>=7.4.3
pytest-qt>=4.3.1
black>=24.1.0
